Removal of all abstract classes for microcontrollers
